Add functionality for running FIJI scripts and macros via widget
Adds some parameters to the FIJI widget that allow a user to specify
the path of a macro or script file that should be run with
FIJI. Additionally, there is a boolean headless parameter that runs
FIJI in headless mode for batch processing using scripts; this should
not be enabled when using macros, which often require use of the GUI.

To enable these to work, the startfiji.sh script is modified to pass
any additional command-line parameters to the chosen `ImageJ-linux64`
executable.

There is also an option to specify a directory containing additional
plugins for FIJI, which can be passed as input from another widget;
this is useful in the case where a user's workflow requires additional
plugins, since it doesn't require modifying the immutable copy of FIJI
inside the Docker container.