Delete bundled Fiji.app in favor of script to build Docker image
Rather than including a copy of FIJI with the repository, this script
will automatically download it, update it, and build the Docker image
that contains it. It will tag the Docker image as
biodepot/fiji:latest; this means to update the copy of FIJI shipped
with the container, it is sufficient to simply run the script and then
push the image to DockerHub.