Remove build_docker_image.sh script as it is no longer needed
The Docker image can now be built just using the Dockerfile; we
previously used the script because our Dockerfile would just add a
local copy of Fiji inside a container, and the script would take care
of updating it. The container build process now involves downloading
and updating Fiji inside the container, so the script is not necessary
anymore.