Add quality of life changes to focal adhesion workflow
Allowed parameter to be passed between Fiji widgets (since if you are
using a custom dataset, we want that to be carried over to the second
Fiji widget as well). Additionally, added a standalone Fiji widget to
just launch Fiji, allowing the user to open Fiji to inspect or modify
the macros and image files midway through a run.